extract monitor rendering and sampling
